Update employees on field clicks instead of watch
React-hook-form doesn't recommend itk, and the useEffect causes an extra
rerender. Instead the logic is moved to whenever a change is registered,
such as changing a field or removing an employee. Whether this will
become slow is up for debate still, but it feels snappier with 3
employees. Might come back and change it